requirements.txt
snowflake-snowpark-python[pandas]
snowflake-cli-labs==0.2.9
streamlit
openai
sqlalchemy
langchain
python-dotenv
snowflake-sqlalchemy